fix(build): add CONFIG_COMPILER_DISABLE_DEFAULT_ERRORS=y to suppress 3rd-party arduino-esp32 warning errors

+++ components/ceepew_hal/ui_manager.h
@@ -95,8 +95,8 @@ typedef struct {
     uint8_t       chat_menu_selected;   /* 0=Read, 1=Write, 2=Check */
     uint8_t       chat_send_confirm_selected; /* 0=Send, 1=Go back */
     char          compose_buffer[CEEPEW_MAX_MSG_CHARS + 4U];  /* Message composition buffer */
-    uint8_t       compose_length;       /* Current message length */
-    uint8_t       compose_cursor;       /* Cursor position in message */
+    uint16_t      compose_length;       /* Current message length */
+    uint16_t      compose_cursor;       /* Cursor position in message */
     uint8_t       keyboard_row;         /* Current keyboard row (0-9 for 6x10 grid) */
     uint8_t       keyboard_col;         /* Current keyboard column (0-5) */
     uint8_t       chat_selected_idx;    /* Currently selected message index in thread */
